Moved b-tagging to JetTaggingUtils and added an option for c-tag
